tidied and documented changes
