+++ app/src/main/java/com/quark/client/navigation/QuarkNavigation.kt
@@ -1,9 +1,15 @@
-package com.quark.client
+package com.quark.client.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.google.firebase.Firebase
+import com.google.firebase.auth.auth
+import com.quark.client.authentication.EmailAuth
+import com.quark.client.pages.Home
+import com.quark.client.pages.Login
+import com.quark.client.pages.SignUp
 
 /**
  * Handles Quark's page navigation
@@ -11,9 +17,11 @@ import androidx.navigation.compose.rememberNavController
 @Composable
 fun QuarkNavigation() {
     val navController = rememberNavController()
+    val auth = EmailAuth(Firebase.auth)
+
     NavHost(navController = navController, startDestination = Screen.Login.route) {
         composable(route = Screen.Login.route) {
-            Login(navController)
+            Login(navController, auth)
         }
         composable(
             route = Screen.Home.route,
@@ -23,7 +31,7 @@ fun QuarkNavigation() {
         composable(
             route = Screen.SignUp.route,
             ) {
-            SignUp(navController)
+            SignUp(navController, auth)
         }
     }
 }
