+++ app/src/main/java/com/quark/client/pages/Login.kt
@@ -1,4 +1,4 @@
-package com.quark.client
+package com.quark.client.pages
 
 import android.app.AlertDialog
 import androidx.compose.foundation.layout.Arrangement
@@ -15,15 +15,15 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import com.google.firebase.Firebase
-import com.google.firebase.auth.auth
 import com.quark.client.authentication.EmailAuth
-
 import android.content.Context
 import androidx.compose.ui.platform.LocalContext
+import com.quark.client.navigation.Screen
+import com.quark.client.authentication.AuthResult
+import com.quark.client.components.showErrorDialog
 
 @Composable
-fun Login(navController: NavController) {
+fun Login(navController: NavController, auth: EmailAuth) {
     val context = LocalContext.current
 
     var email by remember {
@@ -61,12 +61,14 @@ fun Login(navController: NavController) {
             })
         Button(onClick = {
             //Code authenticates user trying to log in.
-            val auth = EmailAuth(Firebase.auth)
+            if(!EmailAuth.validateEmail(email)) {
+                showErrorDialog(context, "$email is formatted incorrectly.\nFormat: Sexton@iu.edu")
+            }
 
             auth.authenticateUser(email, password){
-                if (it == "Authenticated User")
+                if (it == AuthResult.Success) {
                     navController.navigate(Screen.Home.route)
-                else if (it == "Failed"){
+                } else if (it == AuthResult.Failure){
                     showErrorDialog(context, "Please check the entered fields.")
                 }
             }
@@ -80,14 +82,4 @@ fun Login(navController: NavController) {
             Text("Sign Up")
         }
     }
-}
-
-fun showErrorDialog(context: Context, message: String) {
-    AlertDialog.Builder(context)
-        .setTitle("Error")
-        .setMessage(message)
-        .setPositiveButton(android.R.string.ok) { dialog, _ ->
-            dialog.dismiss()
-        }
-        .show()
 }
